requirements.txt
streamlit==1.40.1
folium==0.17.0
paho-mqtt==1.6.1
matplotlib==3.9.2
mysql-connector-python==9.5.0
fpdf==1.7.2
reportlab==4.4.5